chore(tests): updated helpers for cleaner test files
