+++ Pipfile
@@ -9,7 +9,7 @@ httpx = {extras = ["http2"], version = "==0.28.1"}
 [dev-packages]
 pylint = "==4.0.5"
 pytest = "==9.0.2"
-pytest-cov = "==7.0.0"
+pytest-cov = "==7.1.0"
 pytest-httpx = "==0.36.0"
 typing-extensions = "4.15.0"
 switcher-client = {file = ".", editable = true}
