+++ tests/test_client_context.py
@@ -1,8 +1,4 @@
 import pytest
-import time
-
-from typing import Optional
-from pytest_httpx import HTTPXMock
 
 from switcher_client import Client, ContextOptions
 
@@ -24,7 +20,7 @@ def test_context_with_optionals():
 
 def test_context_remote_validation():
     """ Should raise error when missing required fields for remote """
-    
+
     Client.build_context(
         domain='My Domain'
     )
@@ -32,7 +28,7 @@ def test_context_remote_validation():
     # test
     with pytest.raises(ValueError) as excinfo:
         Client.get_switcher().validate() # used by is_on()
-    
+
     assert 'Missing or empty required fields (url, component, api_key)' in str(excinfo.value)
 
 def test_context_get_switcher_from_cache():
@@ -48,13 +44,3 @@ def test_context_get_switcher_from_cache():
 
     assert switcher1 is switcher2
     assert switcher1 is not switcher3
-
-# Helpers
-
-def given_auth(httpx_mock: HTTPXMock, status=200, token: Optional[str]='[token]', exp=int(round(time.time() * 1000))):
-    httpx_mock.add_response(
-        url='https://api.switcherapi.com/criteria/auth',
-        method='POST',
-        status_code=status,
-        json={'token': token, 'exp': exp}
-    )
