+++ tests/test_switcher_local.py
@@ -2,17 +2,19 @@ import json
 
 from time import time
 
+from tests.helpers import given_context
+
 from switcher_client.client import Client, ContextOptions
 from switcher_client.errors import LocalCriteriaError
-from switcher_client.lib.globals.global_context import DEFAULT_REGEX_MAX_BLACKLISTED, DEFAULT_REGEX_MAX_TIME_LIMIT, DEFAULT_RESTRICT_RELAY
 
 async_error = None
+context_options_local = ContextOptions(snapshot_location='tests/snapshots', local=True, logger=True)
 
 def test_local():
     """ Should use local Snapshot to evaluate the switcher """
 
     # given
-    given_context('tests/snapshots')
+    given_context(options=context_options_local)
     snapshot_version = Client.load_snapshot()
 
     switcher = Client.get_switcher()
@@ -25,7 +27,7 @@ def test_local_with_strategy():
     """ Should use local Snapshot to evaluate the switcher with strategy """
 
     # given
-    given_context('tests/snapshots')
+    given_context(options=context_options_local)
     Client.load_snapshot()
 
     switcher = Client.get_switcher()
@@ -40,7 +42,7 @@ def test_local_with_strategy_reset_input():
     """ Should reset strategy inputs """
 
     # given
-    given_context('tests/snapshots')
+    given_context(options=context_options_local)
     Client.load_snapshot()
 
     switcher = Client.get_switcher('FF2FOR2020') \
@@ -51,17 +53,17 @@ def test_local_with_strategy_reset_input():
     assert switcher.is_on()
     switcher.reset_inputs()
     assert switcher.is_on() is False
-    
+
 def test_local_with_strategy_payload():
     """ Should use local Snapshot to evaluate the switcher with payload validation strategy """
 
     # given
-    given_context('tests/snapshots')
+    given_context(options=context_options_local)
     Client.load_snapshot()
 
     switcher = Client.get_switcher()
     payload = {
-        'id': 12345, 
+        'id': 12345,
         'user': {
             'login': 'test_user',
             'role': 'admin'
@@ -72,7 +74,7 @@ def test_local_with_strategy_payload():
     assert switcher \
         .check_payload(json.dumps(payload)) \
         .is_on('FF2FOR2023')
-    
+
     # test (using dict)
     assert switcher \
         .check_payload(payload) \
@@ -82,7 +84,7 @@ def test_local_with_strategy_no_input():
     """ Should return disabled when no input is provided for the strategy """
 
     # given
-    given_context('tests/snapshots')
+    given_context(options=context_options_local)
     Client.load_snapshot()
 
     switcher = Client.get_switcher()
@@ -93,8 +95,11 @@ def test_local_with_strategy_no_input():
 def test_local_with_strategy_no_matching_input():
     """ Should return disabled when no matching input is provided for the strategy """
 
+    options = ContextOptions(**vars(context_options_local))
+    options.regex_max_time_limit = 100
+
     # given
-    given_context(snapshot_location='tests/snapshots', regex_max_time_limit=100)
+    given_context(options=options)
     Client.load_snapshot()
 
     switcher = Client.get_switcher()
@@ -103,18 +108,20 @@ def test_local_with_strategy_no_matching_input():
     assert switcher \
         .check_regex('123') \
         .is_on('FF2FOR2024') is False
-    
+
     # teardown
     Client.clear_resources()
 
 def test_local_with_strategy_redos_input():
     """ Should return disabled when ReDoS input is provided for the strategy """
 
-    # given
+    options = ContextOptions(**vars(context_options_local))
+    options.regex_max_black_list = 1
+    options.regex_max_time_limit = 100
     regex_input = 'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
-    given_context(snapshot_location='tests/snapshots', 
-                  regex_max_black_list=1,
-                  regex_max_time_limit=100)
+
+    # given
+    given_context(options=options)
     Client.load_snapshot()
 
     switcher = Client.get_switcher()
@@ -142,7 +149,7 @@ def test_local_domain_disabled():
     """ Should return disabled when domain is deactivated """
 
     # given
-    given_context('tests/snapshots', environment='default_disabled')
+    given_context(environment='default_disabled', options=context_options_local)
     snapshot_version = Client.load_snapshot()
 
     switcher = Client.get_switcher()
@@ -155,7 +162,7 @@ def test_local_group_disabled():
     """ Should return disabled when group is deactivated """
 
     # given
-    given_context('tests/snapshots')
+    given_context(options=context_options_local)
     snapshot_version = Client.load_snapshot()
 
     switcher = Client.get_switcher()
@@ -168,7 +175,7 @@ def test_local_config_disabled():
     """ Should return disabled when config is deactivated """
 
     # given
-    given_context('tests/snapshots')
+    given_context(options=context_options_local)
     snapshot_version = Client.load_snapshot()
 
     switcher = Client.get_switcher()
@@ -181,7 +188,7 @@ def test_local_strategy_disabled():
     """ Should return disabled when strategy is deactivated """
 
     # given
-    given_context('tests/snapshots')
+    given_context(options=context_options_local)
     snapshot_version = Client.load_snapshot()
 
     switcher = Client.get_switcher()
@@ -194,7 +201,7 @@ def test_local_no_key_found():
     """ Should raise an error when no key is found in the snapshot """
 
     # given
-    given_context('tests/snapshots')
+    given_context(options=context_options_local)
     snapshot_version = Client.load_snapshot()
 
     switcher = Client.get_switcher()
@@ -210,7 +217,7 @@ def test_local_err_check_criteria_default_result():
     """ Should return default result when check criteria fails """
 
     # given
-    given_context('tests/snapshots')
+    given_context(options=context_options_local)
     snapshot_version = Client.load_snapshot()
 
     globals().update(async_error=None)
@@ -219,7 +226,7 @@ def test_local_err_check_criteria_default_result():
 
     # test
     assert snapshot_version == 1
-    
+
     feature = switcher.default_result(True).is_on_with_details('INVALID_KEY')
     assert feature.result is True
     assert feature.reason == 'Default result'
@@ -228,8 +235,11 @@ def test_local_err_check_criteria_default_result():
 def test_local_no_snapshot():
     """ Should raise an error when no snapshot is loaded """
 
+    options = ContextOptions(**vars(context_options_local))
+    options.snapshot_location = 'tests/invalid_location'
+
     # given
-    given_context('tests/invalid_location')
+    given_context(options=options)
     switcher = Client.get_switcher()
 
     # test
@@ -242,7 +252,7 @@ def test_local_retrict_relay():
     """ Should return disabled when Relay is enabled (using default restrict_relay behavior) """
 
     # given
-    given_context('tests/snapshots')
+    given_context(options=context_options_local)
     snapshot_version = Client.load_snapshot()
 
     switcher = Client.get_switcher()
@@ -256,8 +266,11 @@ def test_local_retrict_relay():
 def test_local_restrict_relay_override():
     """ Should return enabled when Relay is enabled but restrict_relay is overridden to False """
 
+    options = ContextOptions(**vars(context_options_local))
+    options.restrict_relay = False
+
     # given
-    given_context('tests/snapshots', restrict_relay=False)
+    given_context(options=options)
     snapshot_version = Client.load_snapshot()
 
     switcher = Client.get_switcher()
@@ -265,22 +278,3 @@ def test_local_restrict_relay_override():
     # test
     assert snapshot_version == 1
     assert switcher.is_on('USECASE103')
-
-# Helpers
-
-def given_context(snapshot_location: str, environment: str = 'default', 
-                  regex_max_black_list = DEFAULT_REGEX_MAX_BLACKLISTED,
-                  regex_max_time_limit = DEFAULT_REGEX_MAX_TIME_LIMIT,
-                  restrict_relay = DEFAULT_RESTRICT_RELAY):
-    Client.build_context(
-        domain='Playground',
-        environment=environment,
-        options=ContextOptions(
-            local=True,
-            logger=True,
-            snapshot_location=snapshot_location,
-            regex_max_black_list=regex_max_black_list,
-            regex_max_time_limit=regex_max_time_limit,
-            restrict_relay=restrict_relay
-        )
-    )
